# swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "service@rizzly.biz.id",
      "name": "RizzyFuzz"
    },
    "title": "E621 API",
    "version": "1.0.0",
    "description": "Example E621 API Documentation"
  },
  "host": "e621api-rizzly.cyclic.app",
  "basePath": "/",
  "paths": {
    "/api/random": {
      "get": {
        "description": "Get a random metadata",
        "parameters": [
          {
            "name": "tags",
            "description": "Comma-separated list of tags",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "error": {
                  "type": "string",
                  "example": "Content or artist Not Found X'3"
                }
              }
            }
          }
        }
      }
    },
    "/api/all": {
      "get": {
        "description": "Get all metadata",
        "parameters": [
          {
            "name": "tags",
            "description": "Comma-separated list of tags",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "404"
                },
                "error": {
                  "type": "string",
                  "example": "Content or artist Not Found X'3"
                }
              }
            }
          }
        }
      }
    }
  }
}